+++ README.md
@@ -13,7 +13,7 @@
   - [x] 숫자 : ACE, 2, 3, 4, 5, 6, 7, 8, 9, J, Q, K
     - ACE만 1(Hard Hand) 또는 11(Soft Hand)로 사용 가능
     - J, Q, K는 10
-  - [ ] 블랙잭에서 사용할 카드는 1벌(52장)만 사용
+  - [x] 블랙잭에서 사용할 카드는 1벌(52장)만 사용
 
 - 카드 발급
   - 딜러
