+++ src/main/java/blackjack/model/Result.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model;
 
 public enum Result {
     WIN,
