+++ src/main/java/blackjack/model/Score.java
@@ -1,6 +1,5 @@
-package blackjack;
+package blackjack.model;
 
-import java.util.List;
 import java.util.Objects;
 
 public class Score {
