+++ src/main/java/blackjack/view/InputView.java
@@ -1,0 +1,5 @@
+package blackjack.view;
+
+public class InputView {
+
+}
