+++ src/test/java/blackjack/model/CardTest.java
@@ -1,9 +1,12 @@
-package blackjack;
+package blackjack.model;
 
-import static blackjack.Rank.*;
-import static blackjack.Suit.*;
+import static blackjack.model.Rank.*;
+import static blackjack.model.Suit.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.Card;
+import blackjack.model.Rank;
+import blackjack.model.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
