+++ src/test/java/blackjack/model/ScoreTest.java
@@ -1,8 +1,8 @@
-package blackjack;
+package blackjack.model;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import blackjack.model.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
